#820 fixed the directory structure
