+++ messages/src/msg_fields/common/mod.rs
@@ -1,0 +1,1 @@
+pub mod attr_value;
