+++ messages/src/msg_fields/mod.rs
@@ -3,4 +3,5 @@
 
 pub mod protocols;
 pub mod traits;
-pub mod common;
+mod common;
+
